backend/requirements.txt
Flask==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.3.1
# Computer Vision - Windows friendly, no compilation needed
opencv-python
opencv-contrib-python
# Core dependencies
numpy
Pillow
python-dotenv
psycopg2-binary
marshmallow
email-validator
# SMS Notifications
twilio==8.10.0
# AI/ML - DeepFace for production face recognition (99.65% accuracy)
tensorflow>=2.13.0
deepface>=0.0.90

# Note: Using DeepFace (Facenet512) for industry-standard face recognition
# TensorFlow backend provides deep learning capabilities
# Face detection models (~100MB) will be downloaded automatically on first use
